+++ src/RoadCaptain.RouteBuilder/IWindowService.cs
@@ -4,11 +4,11 @@ namespace RoadCaptain.RouteBuilder
 {
     public interface IWindowService
     {
-        string ShowOpenFileDialog();
+        string ShowOpenFileDialog(string previousLocation);
         void ShowErrorDialog(string message, Window owner = null);
         void ShowMainWindow();
         void ShowNewVersionDialog(Release release);
-        string ShowSaveFileDialog();
+        string ShowSaveFileDialog(string previousLocation);
         bool ShowDefaultSportSelectionDialog(SportType sport);
         MessageBoxResult ShowSaveRouteDialog();
     }