+++ src/RoadCaptain.UserInterface.Shared/BaseWindowService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Windows;
 using Autofac;
 using Microsoft.Win32;
@@ -17,16 +18,22 @@ namespace RoadCaptain.UserInterface.Shared
 
         protected Window? CurrentWindow { get; private set; }
 
-        public string ShowOpenFileDialog()
+        public string ShowOpenFileDialog(string previousLocation)
         {
+            var initialDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+
+            if (!string.IsNullOrEmpty(previousLocation) && Directory.Exists(previousLocation))
+            {
+                initialDirectory = previousLocation;
+            }
+
             var dialog = new OpenFileDialog
             {
-                RestoreDirectory = true,
                 AddExtension = true,
                 DefaultExt = ".json",
                 Filter = "JSON files (.json)|*.json",
                 Multiselect = false,
-                InitialDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments)
+                InitialDirectory = initialDirectory
             };
 
             var result = ShowDialog(dialog).GetValueOrDefault();